Revert "check whether shm_open accepts a name that contains '+' and more than one '/' (used by Base64 encoding)"
This reverts commit 64bb94b4ce1f37490940f245dfcda86b2c7f2683.
